Upgrade framework to Laravel 11 (with Sanctum 4, Jetstream 5, Pest 3)
Dependency changes:
- laravel/framework ^10.48 -> ^11.0 (resolved 11.55), php ^8.1 -> ^8.2
- laravel/sanctum ^3.2 -> ^4.0; config middleware key renamed to
  validate_csrf_token (table already had expires_at)
- laravel/jetstream ^4.0 -> ^5.0 (published views/actions unchanged:
  drift vs v5 stubs is cosmetic only)
- pestphp/pest-plugin-laravel ^2.2 -> ^3.0 (Pest 3 / PHPUnit 11),
  nunomaduro/collision ^7 -> ^8, laravel/dusk ^8.0 -> ^8.2
- cviebrock/eloquent-sluggable ^10 -> ^11
- rahulhaque/laravel-filepond ^10 -> ^11 (see below)

FilePond v11 port:
- Removed the custom S3FilepondService/S3FilepondController fork
  (v10-era workaround that streamed whole files through PHP to merge
  chunks); v11 ships a native S3 multipart driver selected from the
  temp_disk driver
- Migration renames fileponds.mimetypes -> mimetype and adds metadata,
  upload_id, upload_tags columns required by the new driver
- JS chunk-size floors raised to 5 MiB (S3 multipart minimum)

Carbon 3 sweep (arrives with Laravel 11):
- diffIn* now returns signed floats; all 73 call sites across app/ and
  views updated to preserve Carbon 2 absolute-integer semantics via
  explicit absolute arg + (int) casts

Laravel 11 compatibility fixes:
- Blade component-tag attributes can no longer contain @if directives
  or multi-line match() expressions (compiled to invalid PHP under the
  Livewire stack); hoisted into @php blocks in overview-card,
  manage-portfolio-items, and payouts views
- portfolio_items migration now drops the FK before dropColumn on all
  drivers; L11's native SQLite DROP COLUMN rejects FK-referenced
  columns and its schema grammar supports dropForeign on SQLite

Full suite green: 1454 passed, 62 skipped (baseline 1442/62 on L10).

+++ resources/views/components/client-project/compact-header.blade.php
@@ -49,7 +49,7 @@
     $daysRemaining = null;
     if ($project->deadline) {
         $deadline = \Carbon\Carbon::parse($project->deadline);
-        $daysRemaining = now()->diffInDays($deadline, false);
+        $daysRemaining = (int) now()->diffInDays($deadline, false);
     }
 @endphp
